==> R_scripts/11_ScPred_3D.Rmd <==
---
title: "11_ScPred_3D"
author: "BH"
date: '2022-04-22'
output:
 pdf_document:
    toc: yes
---
```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

```{r load_packages}
library("dplyr")
library("Seurat")
library("patchwork")
library("readxl")
library("tidyseurat")
library("ggplot2")
library("tidyverse")
library("RCurl")
library("cowplot")
library("sctransform")
library("scPred")
library("Seurat")
library("magrittr")
library("caret")
library("dplyr")
library("janitor")
library("factoextra")
library("reshape2")

options(future.globals.maxSize = 12000 * 1024^2)

```
# Filtering Seurat to select just 3D cells for clustering 
```{r}
custom_theme <- readRDS(file = "custom_theme.rds")
hdfseurat_filt <- readRDS(file = "hdfseurat_filt.rds")

hdfseurat_filt_3D <- subset(hdfseurat_filt, Sample == "DS3D" | Sample == "EP3D")
```

# Plotting top 10 variable features in the 3D cells
```{r}
#Selecting highly variable features - high cell to cell variation
hdfseurat_filt_3D <- FindVariableFeatures(hdfseurat_filt_3D, selection.method ="vst", nfeatures=2000)
#Identify the 10 most highly variable genes
top10 <- head(VariableFeatures(hdfseurat_filt_3D), 10)
top10
plot3 <- VariableFeaturePlot(hdfseurat_filt_3D)
#Plot variable features with labels
plot4 <- LabelPoints(plot=plot3, points = top10, repel=TRUE)
plot4

ggsave(
  "HDF_3D_variablefeatures.tiff",
  plot = last_plot(),
  device = "tiff",
  path = "../Output/HDF_Data_Analysis"
  )
```

# Scaling the data before dimensionality reduction. Each gene is centred to a mean of 0, and scaled by the standard deviation of each gene.
```{r}
all.genes <- rownames(hdfseurat_filt_3D)
hdfseurat_filt_3D <- ScaleData(hdfseurat_filt_3D, features = all.genes)

#Performing a PCA
hdfseurat_filt_3D <- RunPCA(hdfseurat_filt_3D, features=VariableFeatures(object = hdfseurat_filt_3D))

#Plotting the PCA
PCA_no_label <- DimPlot(hdfseurat_filt_3D, reduction = "pca", seed = 123, combine=TRUE) + NoLegend()
PCA_Sample_label <- DimPlot(hdfseurat_filt_3D, reduction = "pca", seed = 123, group.by = "Sample")

PCA_no_label
PCA_Sample_label
```

## Heatmaps to investigate the genes influencing the top PCs
```{r}
DimHeatmap(hdfseurat_filt_3D, dims=1, cells=500, balanced=TRUE)
DimHeatmap(hdfseurat_filt_3D, dims=1:15, cells=500, balanced=TRUE)
ElbowPlot(hdfseurat_filt_3D) 

#Suggests top 10 PCs explain most of the variance. Test this with jackstraw plot
#hdfseurat_jackstraw <- JackStraw(hdfseurat_filt_3D, num.replicate = 100)
#hdfseurat_jackstraw <- ScoreJackStraw(hdfseurat_jackstraw, dims = 1:20)
#JackStrawPlot(hdfseurat_jackstraw, dims = 1:15)
```

## Clustering the cells using KNN
```{r}
#Use K-nearest neighbours based on PCA distance in FindNeighbours() function based on previously definited dimensions (10) from elbow plot
hdfseurat_filt_3D <- FindNeighbors(hdfseurat_filt_3D, dims =1:10)

#group cells using findclusters function
hdfseurat_filt_3D <- FindClusters(hdfseurat_filt_3D, resolution = 0.6)

#Look at cluster ID's of first 5 cells
head(Idents(hdfseurat_filt_3D), 5)
```

## Plot UMAP
```{r}
#Running a UMAP
hdfseurat_filt_3D <- RunUMAP(hdfseurat_filt_3D, dims=1:10)

#Plotting UMAP with cluster IDs
DimPlot(hdfseurat_filt_3D, reduction = "umap", seed = 123, cols = c("#9b5fe0", "#16a5d8", "#60dbe8", "#8bd346", "#efdf48", "#f9a52c", "#d64e12" )) +
  custom_theme(base_size=20)
ggsave(
  "UMAP_HDF_3D.tiff",
  plot = last_plot(),
  device = "tiff",
  path = "../Output/HDF_Data_Analysis"
  )

#Colouring above UMAP plot with Sample type
DimPlot(hdfseurat_filt_3D, reduction = "umap",group.by = "Sample", seed=123, cols=c( "#24B3A8", "#FF92A5")) +
  custom_theme(base_size=20)
ggsave(
  "UMAP_HDF_3D_samplecolour.tiff",
  plot = last_plot(),
  device = "tiff",
  path = "../Output/HDF_Data_Analysis"
  )
```
# Finding cluster biomarkers
```{r}
#Find markers for every cluster compared to all remaining cells, report only the positive ones
hdfseurat.markers <- FindAllMarkers(hdfseurat_filt_3D, only.pos=FALSE, min.pct=0.25, logfc.threshold=0.25)

write.csv(hdfseurat.markers,"../Output/HDF_Data_Analysis/HDF3D_clustermarkers_posandneg.csv")
```

# Building 3D ML model using ScPred

## Seperating dataset into training and testing. 80% training, and 20% testing
```{r}
#Extracting just 3D senescent and proliferative cells
new_col <- 1:769

hdfseurat_filt_3D@meta.data <- cbind(hdfseurat_filt_3D@meta.data, new_col)

hdfseurat_filt_3D@meta.data$new_col <- 1:nrow(hdfseurat_filt_3D@meta.data)
train <- hdfseurat_filt_3D@meta.data %>% dplyr::sample_frac(.8)
test  <- dplyr::anti_join(hdfseurat_filt_3D@meta.data, train, by = 'new_col')

test <- test$new_col
train <- train$new_col

hdfseurat_filt_3D@meta.data$new_col[hdfseurat_filt_3D@meta.data$new_col %in% test] <- "Test"
hdfseurat_filt_3D@meta.data$new_col[hdfseurat_filt_3D@meta.data$new_col %in% train] <- "Train"

obj_list <- SplitObject(hdfseurat_filt_3D, split.by = "new_col")
```

## Normalising, scaling and running PCA on training and testing data
```{r}
training <- obj_list$Train %>% 
  NormalizeData() %>% 
  FindVariableFeatures() %>% 
  ScaleData() %>% 
  RunPCA() %>% 
  RunUMAP(dims = 1:10)

DimPlot(training, group.by = "Sample", label = FALSE, repel = TRUE, seed =123)

ggsave(
  "UMAP_3D_training.tiff",
  plot = last_plot(),
  device = "tiff",
  path = "../Output/HDF_ScPred"
  )

testing <- obj_list$Test%>%
  NormalizeData() %>%
  FindVariableFeatures() %>% 
  ScaleData() %>% 
  RunPCA() %>% 
  RunUMAP(dims = 1:10)

DimPlot(testing, group.by = "Sample", label = FALSE, repel=TRUE, seed = 123)

ggsave(
  "UMAP_3D_testing.tiff",
  plot = last_plot(),
  device = "tiff",
  path = "../Output/HDF_ScPred"
  )

##Save training and testing datasets
saveRDS(training, file = "../Output/Seurat_Objects/hdfseuratfiltered_3Dtraining.rds")
saveRDS(testing, file =  "../Output/Seurat_Objects/hdfseuratfiltered_3Dtesting_EPESDS.rds" )

## Note - every time you run the script you will get slightly different results. This is because the training and testing splits are random. If you want to exactly replicate your results, move the training and testing files into 'Input/Seurat_Objects' and load them in each time.

training <- readRDS("../Inputfiles/Seurat_Objects/hdfseuratfiltered_3Dtraining.rds")
testing <- readRDS("../Inputfiles/Seurat_Objects/hdfseuratfiltered_3Dtesting.rds")
```

```{r, results = FALSE}
#Using the training dataset to create and evaluate different models - no PC filters

hdfseurat_scPred <- getFeatureSpace(training, "Sample")


#This model is the svm Radial model
hdfseurat_svmradial3D <- trainModel(hdfseurat_scPred)
get_probabilities(hdfseurat_svmradial3D) %>% head()
get_scpred(hdfseurat_svmradial3D)
hdfseurat_svmradial3D
get_classifiers(hdfseurat_svmradial3D)


#This model is the mda model
hdf_mda3D <- trainModel(hdfseurat_scPred, model = "mda")
get_probabilities(hdf_mda3D) %>% head()
get_scpred(hdf_mda3D)
hdf_mda3D

#This model is logistic regression
hdf_glm3D <- trainModel(hdfseurat_scPred, model = "glm")
get_probabilities(hdf_glm3D)%>% head()
get_scpred(hdf_glm3D)
hdf_glm3D



#This model is knn
hdf_knn3D <- trainModel(hdfseurat_scPred, model = "knn")
get_probabilities(hdf_knn3D)%>% head()
get_scpred(hdf_knn3D)
hdf_knn3D
```

## Decided to take SVM and MDA forward for further analysis due to increased sens and spec. 
Testing on test dataset

```{r}

#SVM testing 
query <- NormalizeData(testing)
query <- scPredict(query, hdfseurat_svmradial3D, recompute_alignment = FALSE)

#Plotting UMAP of predicted senescent cells
query <- RunUMAP(query, reduction = "scpred", dims = 1:10)
DimPlot(query, group.by = "scpred_prediction", label = FALSE, repel = TRUE, seed = 123, cols = c("darkblue", "deeppink3", "antiquewhite4"))
ggsave(
  "UMAP_scpred_testing3D_SVM.tiff",
  plot = last_plot(),
  device = "tiff",
  path = "../Output/HDF_Data_ScPred/3D_SVM"
  )
DimPlot(query, group.by = "Sample", label = FALSE, repel = TRUE, seed = 123)
ggsave(
  "UMAP_scpred_testing3D_SAMPLE.tiff",
  plot = last_plot(),
  device = "tiff",
  path = "../Output/HDF_Data_ScPred/3D_SVM"
  )
FeaturePlot(query, c("scpred_DS3D", "scpred_EP3D"))
```

##Confusion matrix plot
```{r}
query_df <- query@meta.data
query_df <- query_df %>%
  mutate(Status = case_when(query_df$Sample=="DS3D" & query_df$scpred_prediction=="DS3D" ~ "DS3DT"))

ds3Dt <- nrow(query_df[!is.na(query_df$Status),])

query_df <- query_df %>%
  mutate(Status = case_when(query_df$Sample=="DS3D" & query_df$scpred_prediction=="EP3D" ~ "EP3DF"))

ep3Df <- nrow(query_df[!is.na(query_df$Status),])

query_df <- query_df %>%
  mutate(Status = case_when(query_df$Sample=="EP3D" & query_df$scpred_prediction=="EP3D" ~ "EP3DT"))

ep3Dt <- nrow(query_df[!is.na(query_df$Status),])


query_df <- query_df %>%
  mutate(Status = case_when(query_df$Sample=="EP3D" & query_df$scpred_prediction=="DS3D" ~ "DS3DF"))

ds3Df <- nrow(query_df[!is.na(query_df$Status),])

unassigned <- length(which(query_df$scpred_prediction == "unassigned"))

#Plotting into confusion matrix

table <- matrix(c(ds3Dt, ep3Df, ds3Df, ep3Dt),ncol=2,byrow=TRUE)
colnames(table) <- c("DS", "EP")
rownames(table) <- c("DS", "EP")
table <- as.table(table)
table

perc <- prop.table(table, 1)*100
perc

perc <- round(perc, digits=2)

cm <- confusionMatrix(perc)


# Drawing confusion matrix into visual table

draw_confusion_matrix <- function(cm) {

  layout(matrix(c(1,1,2)))
  par(mar=c(2,2,2,2))
  plot(c(100, 345), c(300, 450), type = "n", xlab="", ylab="", xaxt='n', yaxt='n')
  title('SVM Confusion Matrix (%)', cex.main=2)

  # create the matrix 
  rect(150, 430, 240, 370, col='#32CD32')
  text(195, 440, 'DS', cex=1.2)
  rect(250, 430, 340, 370, col='#FF4500')
  text(295, 440, 'EP', cex=1.2)
  text(125, 370, 'Predicted', cex=1.3, srt=90, font=2)
  text(245, 450, 'Actual', cex=1.3, font=2)
  rect(150, 305, 240, 365, col='#FF4500')
  rect(250, 305, 340, 365, col='#32CD32')
  text(140, 400, 'DS', cex=1.2, srt=90)
  text(140, 335, 'EP', cex=1.2, srt=90)

  # add in the cm results 
  res <- as.numeric(cm$table)
  text(195, 400, res[1], cex=1.6, font=2, col='white')
  text(195, 335, res[2], cex=1.6, font=2, col='white')
  text(295, 400, res[3], cex=1.6, font=2, col='white')
  text(295, 335, res[4], cex=1.6, font=2, col='white')
  text(110, 300, 'Unassigned =', cex=0.9, font=1, col='black')
  text(130, 300, unassigned, cex=0.9, font=1, col='black')


}  

draw_confusion_matrix(cm)

```

#Testing MDA model
```{r}
#SVM testing 
query <- NormalizeData(testing)
query <- scPredict(query, hdf_mda3D, recompute_alignment = FALSE)

#Plotting UMAP of predicted senescent cells
query <- RunUMAP(query, reduction = "scpred", dims = 1:10)
DimPlot(query, group.by = "scpred_prediction", label = FALSE, repel = TRUE, seed = 123, cols = c("darkblue", "deeppink3")) +
  custom_theme(base_size=20)
ggsave(
  "UMAP_scpred_testing3D.tiff",
  plot = last_plot(),
  device = "tiff",
  path = "../Output/HDF_Data_ScPred/3D_MDA"
  )
DimPlot(query, group.by = "Sample", label = FALSE, repel = TRUE, seed = 123) +
  custom_theme(base_size=20)
ggsave(
  "UMAP_scpred_testing3D_SAMPLE.tiff",
  plot = last_plot(),
  device = "tiff",
  path = "../Output/HDF_Data_ScPred/3D_MDA"
  )
FeaturePlot(query, c("scpred_DS3D", "scpred_EP3D"))
```

##Confusion matrix plot
```{r}
query_df <- query@meta.data
query_df <- query_df %>%
  mutate(Status = case_when(query_df$Sample=="DS3D" & query_df$scpred_prediction=="DS3D" ~ "DS3DT"))

ds3Dt <- nrow(query_df[!is.na(query_df$Status),])

query_df <- query_df %>%
  mutate(Status = case_when(query_df$Sample=="DS3D" & query_df$scpred_prediction=="EP3D" ~ "EP3DF"))

ep3Df <- nrow(query_df[!is.na(query_df$Status),])

query_df <- query_df %>%
  mutate(Status = case_when(query_df$Sample=="EP3D" & query_df$scpred_prediction=="EP3D" ~ "EP3DT"))

ep3Dt <- nrow(query_df[!is.na(query_df$Status),])


query_df <- query_df %>%
  mutate(Status = case_when(query_df$Sample=="EP3D" & query_df$scpred_prediction=="DS3D" ~ "DS3DF"))

ds3Df <- nrow(query_df[!is.na(query_df$Status),])

unassigned <- length(which(query_df$scpred_prediction == "unassigned"))

#Plotting into confusion matrix

table <- matrix(c(ds3Dt, ep3Df, ds3Df, ep3Dt),ncol=2,byrow=TRUE)
colnames(table) <- c("DS", "EP")
rownames(table) <- c("DS", "EP")
table <- as.table(table)
table

perc <- prop.table(table, 1)*100
perc

perc <- round(perc, digits=2)

cm <- confusionMatrix(perc)

# Drawing confusion matrix into visual table

draw_confusion_matrix <- function(cm, output_file) {

  draw_matrix_plot <- function() {

  layout(matrix(c(1,1,2)))
  par(mar=c(2,2,2,2))
  plot(c(100, 345), c(300, 450), type = "n", xlab="", ylab="", xaxt='n', yaxt='n')
  title('MDA Confusion Matrix (%)', cex.main=2)

  # create the matrix 
  rect(150, 430, 240, 370, col='#32CD32')
  text(195, 440, 'DS', cex=1.2)
  rect(250, 430, 340, 370, col='#FF4500')
  text(295, 440, 'EP', cex=1.2)
  text(125, 370, 'Predicted', cex=1.3, srt=90, font=2)
  text(245, 450, 'Actual', cex=1.3, font=2)
  rect(150, 305, 240, 365, col='#FF4500')
  rect(250, 305, 340, 365, col='#32CD32')
  text(140, 400, 'DS', cex=1.2, srt=90)
  text(140, 335, 'EP', cex=1.2, srt=90)

  # add in the cm results 
  res <- as.numeric(cm$table)
  text(195, 400, res[1], cex=1.6, font=2, col='white')
  text(195, 335, res[2], cex=1.6, font=2, col='white')
  text(295, 400, res[3], cex=1.6, font=2, col='white')
  text(295, 335, res[4], cex=1.6, font=2, col='white')
  text(110, 300, 'Unassigned =', cex=0.9, font=1, col='black')
  text(130, 300, unassigned, cex=0.9, font=1, col='black')

}  

  # Call the function to draw the matrix plot
  draw_matrix_plot()

  # Save the plot as a PNG file
  png(filename = output_file, width = 800, height = 600)  # Set appropriate width and height
  draw_matrix_plot()
  dev.off()  # Turn off the PNG device
}


your_confusion_matrix <- cm
output_file_path <- "../Output/HDF_ScPred/scPred3D_MDA_CM.png"
draw_confusion_matrix(your_confusion_matrix, output_file_path)

```

```{r}
saveRDS(hdf_mda3D, file = "hdf_mda3D.rds")
saveRDS(hdfseurat_filt_3D, file="hdfseurat_filt_3D.rds")

```

# Generating a correlation matrix of a panel of selected senescence genes - EP cells
```{r}
# Subset the Seurat object based on scpred_prediction
filtered_seurat_EP <- subset(hdfseurat_filt_3D, subset = Sample == "EP3D")

# Extract gene counts for filtered cells
gene_counts <- filtered_seurat_EP@assays$RNA@data

# Define gene list 
gene_list <- c("CDKN1A", "CDKN2A", "IGFBP3", "SERPINE1", "IGF1", "CXCL12", "IL6", "MMP3", "TP53")

# Check if gene names in gene_list are present in row names (genes)
missing_genes <- gene_list[!gene_list %in% rownames(gene_counts)]
if (length(missing_genes) > 0) {
  stop(paste("Genes missing in gene counts:", paste(missing_genes, collapse = ", ")))
}

# Filter gene counts based on the gene list
gene_counts <- gene_counts[gene_list, ]
gene_counts <- as.matrix(gene_counts)
# Transpose gene counts matrix
gene_counts_transposed <- t(gene_counts)
gene_counts_transposed <- as.matrix(gene_counts_transposed)

# Calculate correlation matrix
correlation_matrix <- cor(gene_counts_transposed)

# Print correlation matrix
print(correlation_matrix)

dendrogram_row <- as.dendrogram(hclust(dist(correlation_matrix), method = "ward.D2"))
dendrogram_col <- as.dendrogram(hclust(dist(t(correlation_matrix)), method = "ward.D2"))

# Define breaks and color palette
breaks <- unique(c(seq(-1,-0.0001,length=50),seq(-0.00001,0.01,length=50), seq(0.011,1,length=50)))
my_palette <- colorRampPalette(c("blue","white", "white", "red"))(length(breaks)-1)

library("gplots")
png(filename = "../Output/HDF_Data_Analysis/Markercorrelation_EP3D.png", width = 800, height = 600)  # Set appropriate width and height

heatmap <- heatmap.2(correlation_matrix, 
                     col = my_palette, 
                     breaks = breaks, 
                     key = TRUE, 
                     trace = "none", 
                     main = "Marker Correlation_EP_3D", 
                     key.title = "Correlation",   
                     cexRow = 0.6, 
                     cexCol = 0.6,    
                     distfun = function(x) dist(x, method = "euclidean"),
                     hclust = function(x) hclust(x, method = "ward.D2"), 
                     symkey = FALSE, 
                     Rowv = dendrogram_row,
                     Colv = dendrogram_col,
                     dendrogram = "both",
                     density.info = c("none")) +
  custom_theme(base_size=20)

dev.off()

print(heatmap)
```
# Generating a correlation matrix of a panel of selected senescence genes - DS cells
```{r}
# Subset the Seurat object based on scpred_prediction
filtered_seurat_EP <- subset(hdfseurat_filt_3D, subset = Sample == "DS3D")

# Extract gene counts for filtered cells
gene_counts <- filtered_seurat_EP@assays$RNA@data

# Define gene list 
gene_list <- c("CDKN1A", "CDKN2A", "IGFBP3", "SERPINE1", "IGF1", "CXCL12", "IL6", "MMP3", "TP53")

# Check if gene names in gene_list are present in row names (genes)
missing_genes <- gene_list[!gene_list %in% rownames(gene_counts)]
if (length(missing_genes) > 0) {
  stop(paste("Genes missing in gene counts:", paste(missing_genes, collapse = ", ")))
}

# Filter gene counts based on the gene list
gene_counts <- gene_counts[gene_list, ]
gene_counts <- as.matrix(gene_counts)
# Transpose gene counts matrix
gene_counts_transposed <- t(gene_counts)
gene_counts_transposed <- as.matrix(gene_counts_transposed)

# Calculate correlation matrix
correlation_matrix <- cor(gene_counts_transposed)

# Print correlation matrix
print(correlation_matrix)

# Define breaks and color palette
breaks <- unique(c(seq(-1,-0.0001,length=50),seq(-0.00001,0.01,length=50), seq(0.011,1,length=50)))
my_palette <- colorRampPalette(c("blue","white", "white", "red"))(length(breaks)-1)

library("gplots")
png(filename = "../Output/HDF_Data_Analysis/Markercorrelation_DS3D.png", width = 800, height = 600)  # Set appropriate width and height

heatmap <- heatmap.2(correlation_matrix, 
                     col = my_palette, 
                     breaks = breaks, 
                     key = TRUE, 
                     trace = "none", 
                     main = "Marker Correlation_DS_3D", 
                     key.title = "Correlation",   
                     cexRow = 0.6, 
                     cexCol = 0.6,    
                     distfun = function(x) dist(x, method = "euclidean"),
                     hclust = function(x) hclust(x, method = "ward.D2"), 
                     symkey = FALSE, 
                     Rowv = dendrogram_row,
                     Colv = dendrogram_col,
                     dendrogram = "both",
                     density.info = c("none")) +
  custom_theme(base_size=20)

dev.off()

print(heatmap)
```

### Session information
```{r session_info_99}
sessionInfo()
```

This document was processed on: `r Sys.Date()`.
