==> R_scripts/15_matrisome.Rmd <==
---
title: "15_Matrisome"
author: "BH"
date: "2024-02-07"
output:
 pdf_document:
    toc: yes
---
```{r setup, include=FALSE} 
knitr::opts_chunk$set(warning = FALSE, message = FALSE) 
```

```{r load_packages}
library("dplyr")
library("Seurat")
library("patchwork")
library("readxl")
library("tidyseurat")
library("ggplot2")
library("tidyverse")
library("RCurl")
library("cowplot")
library("sctransform")
library("scPred")
library("Seurat")
library("magrittr")
library("caret")
library("dplyr")
library("janitor")
library("factoextra")
library("reshape2")

options(future.globals.maxSize = 12000 * 1024^2)

```
# Matrisome genes were imported from the Matrisome project: https://sites.google.com/uic.edu/matrisome/home?authuser=0
The genes imported were the NABA_MATRISOME human gene set. 

```{r}
custom_theme <- readRDS(file = "custom_theme.rds")
Tabib_fibs2 <- readRDS(file="Tabib_fibs2.rds")
Sole_fibs2 <- readRDS(file="Sole_fibs2.rds")
hdfseurat_filt_2D <- readRDS(file = "../Output/Seurat_Objects/hdfseuratfiltered_2Dcells.rds")
hdfseurat_filt_3D <- readRDS(file="hdfseurat_filt_3D.rds")
matrisome <- read.csv ("../Inputfiles/matrisome.csv", header = FALSE)

```

```{r}
#Extracting gene names from the matrisome list
matrisome_genes <- matrisome$V1

# Find common genes between 'matrisome_genes' and the gene names in the Seurat object
common_genes <- intersect(rownames(hdfseurat_filt_2D), matrisome_genes)

# Check if common genes were found
if (length(common_genes) == 0) {
  stop("None of the features in 'matrisome' are present in the Seurat object.")
}

# Subset the Seurat object to include only the common genes
hdfseurat_filt_2D_matrisome <- hdfseurat_filt_2D@assays[["RNA"]]@counts
df <- as.data.frame(hdfseurat_filt_2D_matrisome)

filtered_df <- df[rownames(df) %in% common_genes, ]

# Calculate row medians
row_medians <- apply(filtered_df, 1, median)

# Create a dataframe with only row medians
median_df_2D <- data.frame("mediancounts_2D" = row_medians, row.names = rownames(filtered_df))

```


```{r}
matrisome_genes <- matrisome$V1

# Find common genes between 'matrisome_genes' and the gene names in the Seurat object
common_genes <- intersect(rownames(hdfseurat_filt_3D), matrisome_genes)

# Check if common genes were found
if (length(common_genes) == 0) {
  stop("None of the features in 'matrisome' are present in the Seurat object.")
}

# Subset the Seurat object to include only the common genes
hdfseurat_filt_3D_matrisome <- hdfseurat_filt_3D@assays[["RNA"]]@counts
df <- as.data.frame(hdfseurat_filt_3D_matrisome)

filtered_df <- df[rownames(df) %in% common_genes, ]

# Calculate row medians
row_medians <- apply(filtered_df, 1, median)

# Create a dataframe with only row medians
median_df_3D <- data.frame("mediancounts_3D" = row_medians, row.names = rownames(filtered_df))


```

```{r}
matrisome_genes <- matrisome$V1

# Find common genes between 'matrisome_genes' and the gene names in the Seurat object
common_genes <- intersect(rownames(Tabib_fibs2), matrisome_genes)

# Check if common genes were found
if (length(common_genes) == 0) {
  stop("None of the features in 'matrisome' are present in the Seurat object.")
}

# Subset the Seurat object to include only the common genes
Tabib_fibs2_matrisome <- Tabib_fibs2@assays[["RNA"]]@counts
df <- as.data.frame(Tabib_fibs2_matrisome)

filtered_df <- df[rownames(df) %in% common_genes, ]

# Calculate row medians
row_medians <- apply(filtered_df, 1, median)

# Create a dataframe with only row medians
median_df_tabib <- data.frame("mediancounts_Tabib" = row_medians, row.names = rownames(filtered_df))


```

```{r}
matrisome_genes <- matrisome$V1

# Find common genes between 'matrisome_genes' and the gene names in the Seurat object
common_genes <- intersect(rownames(Sole_fibs2), matrisome_genes)

# Check if common genes were found
if (length(common_genes) == 0) {
  stop("None of the features in 'matrisome' are present in the Seurat object.")
}

# Subset the Seurat object to include only the common genes
Sole_fibs2_matrisome <- Sole_fibs2@assays[["RNA"]]@counts
df <- as.data.frame(Sole_fibs2_matrisome)

filtered_df <- df[rownames(df) %in% common_genes, ]

# Calculate row medians
row_medians <- apply(filtered_df, 1, median)

# Create a dataframe with only row medians
median_df_sole <- data.frame("mediancounts_Sole" = row_medians, row.names = rownames(filtered_df))


```

```{r}

# Check common row names
common_row_names <- Reduce(intersect, list(rownames(median_df_2D), rownames(median_df_3D), rownames(median_df_sole), rownames(median_df_tabib)))


# Subset the dataframes to include only the common row names and the column representing median counts
median_df_2D_subset <- median_df_2D[common_row_names, "mediancounts_2D", drop = FALSE]
median_df_3D_subset <- median_df_3D[common_row_names, "mediancounts_3D", drop = FALSE]
median_df_sole_subset <- median_df_sole[common_row_names, "mediancounts_Sole", drop = FALSE]
median_df_tabib_subset <- median_df_tabib[common_row_names, "mediancounts_Tabib", drop = FALSE]

# Merge the dataframes using cbind
merged_df <- cbind(median_df_2D_subset, median_df_3D_subset, median_df_sole_subset, median_df_tabib_subset)

# Rename the row names column
rownames(merged_df) <- common_row_names



```

# Generating a correlation matrix of median matrisome gene counts
```{r}
# Load required libraries
library(gplots)

# Extract numeric matrix from merged_df, excluding the 'genes' column
numeric_matrix <- as.matrix(merged_df)

# Create breaks and define palette
breaks <- unique(c(seq(-10,-0.011,length=100), seq(-0.011,0.01,length=100), seq(0.011,10,length=100)))
my_palette <- colorRampPalette(c("red","white", "white", "blue"))(length(breaks)-1)

# Create heatmap using heatmap.2 with adjusted x-axis label size
heatmap.2(numeric_matrix, col=my_palette, breaks = breaks, key=TRUE, trace="none", main="Matrisome", key.title = 1, cexRow=0.1, cexCol=0.8, distfun = function(x) dist(x, method = "euclidean"), hclust = function(x) hclust(x, method = "ward.D2"), symkey = FALSE, density.info = "none", cex.axis = 5)


```

```{r}
# Remove rows where all values are 0
merged_df_nozero <- merged_df[rowSums(merged_df != 0) != 0, ]
# Load required libraries
library(gplots)

# Extract numeric matrix from merged_df, excluding the 'genes' column
numeric_matrix <- as.matrix(merged_df_nozero)

# Create breaks and define palette
breaks <- unique(c(seq(-0.011,0.01,length=100), seq(0.011,20,length=100)))
my_palette <- colorRampPalette(c("white", "white", "blue"))(length(breaks)-1)


png(filename = "../Output/Matrisome/Matrisome_nozero.png", width = 800, height = 600)  # Set appropriate width and height

# Create heatmap using heatmap.2 with adjusted x-axis label size
heatmap.2(numeric_matrix, col=my_palette, breaks = breaks, key=TRUE, trace="none", main="Matrisome", key.title = "Gene count", cexRow=0.1, cexCol=0.5, distfun = function(x) dist(x, method = "euclidean"), hclust = function(x) hclust(x, method = "ward.D2"), symkey = FALSE, density.info = "none", cex.axis = 5) 

dev.off()

heatmap.2(numeric_matrix, col=my_palette, breaks = breaks, key=TRUE, trace="none", main="Matrisome", key.title = "Gene count", cexRow=0.1, cexCol=0.5, distfun = function(x) dist(x, method = "euclidean"), hclust = function(x) hclust(x, method = "ward.D2"), symkey = FALSE, density.info = "none", cex.axis = 5) 
```
```{r}
merged_df_nozero <- rownames(merged_df_nozero)
merged_df_nozero <- as.data.frame(merged_df_nozero)
write.csv(merged_df_nozero, file = "../Output/Matrisome/merged_df_nozero.csv", row.names = TRUE)
```

### Session information
```{r session_info_99}
sessionInfo()
```
