==> R_scripts/99_Run_All.Rmd <==
---
title: "99_Run_All"
author: ""
date: "`r format(Sys.time(), '%d %B %Y')`"
output:
  rmarkdown::html_document:
    toc: TRUE
---


# Install/Load dependencies
```{r, echo=TRUE, message=FALSE, Loading_Dependencies_99}
# Load dependencies:
library(here)
library(rmarkdown)
```

```{r Run_all, include=FALSE}
files_in_r_to_run <- 
  c("1_HDF_data.Rmd","2_HDF_filtering_normalisation.Rmd", "3_HDF_2DCells.Rmd", "4_ScPred_2D.Rmd", "5_ScPred2d_Chan.Rmd", "6_ScPred2D_Chan_EPESDS.Rmd", "7_Teo.Rmd", "8_Tabib.Rmd", "9_soleboldo.Rmd", "10_soleboldo_tabib_combined2D.rmd", "11_ScPred_3D.Rmd", "12_3Dpred_Tabib.Rmd", "13_3Dpred_SoleBoldo.Rmd", "14_soleboldo_tabib_combined3D.Rmd", "15_matrisome.Rmd", "16_Ganier.Rmd", "17_Ganier2.Rmd", "18_Ganier3.Rmd", "19_alldermal_celltypes.Rmd", "20_allskin_ganier.Rmd", "21_lungs.Rmd", "22_lungs.Rmd")

for(i1 in 1:length(files_in_r_to_run)){
  
  rmarkdown::render(here("R_scripts", files_in_r_to_run[i1]),
                    output_format = 
                      html_document(html_preview = TRUE, toc = TRUE),
                    output_dir = here("Output"))
}

rmarkdown::render(here("README.Rmd"),
                  output_format =  html_document(html_preview = TRUE, toc = TRUE,keep_md=TRUE),
                  output_dir = here())

```



### Session information
```{r session_info_99}
sessionInfo()
```

This document was processed on: `r Sys.Date()`.
